Continue setting up custom notifications types
