+++ packages/liveblocks-core/src/index.ts
@@ -183,9 +183,12 @@ export type {
 } from "./types/CommentReaction";
 export type { Immutable } from "./types/Immutable";
 export type {
+  InboxNotificationCustomData,
+  InboxNotificationCustomDataPlain,
   InboxNotificationData,
   InboxNotificationDataPlain,
   InboxNotificationThreadData,
+  InboxNotificationThreadDataPlain,
 } from "./types/InboxNotificationData";
 export type { InboxNotificationDeleteInfo } from "./types/InboxNotificationDeleteInfo";
 export type {
