+++ packages/liveblocks-react-comments/src/components/InboxNotification.tsx
@@ -1,17 +1,11 @@
 "use client";
 
 import type {
-  BaseMetadata,
-  CommentData,
+  InboxNotificationCustomData,
   InboxNotificationData,
   InboxNotificationThreadData,
-  ThreadData,
 } from "@liveblocks/core";
-import {
-  assertNever,
-  getMentionedIdsFromCommentBody,
-  kInternal,
-} from "@liveblocks/core";
+import { assertNever, kInternal } from "@liveblocks/core";
 import { useLiveblocksContextBundle } from "@liveblocks/react";
 import { TooltipProvider } from "@radix-ui/react-tooltip";
 import type {
@@ -34,50 +28,30 @@ import type {
   InboxNotificationOverrides,
 } from "../overrides";
 import { useOverrides } from "../overrides";
-import * as CommentPrimitive from "../primitives/Comment";
 import { Timestamp } from "../primitives/Timestamp";
 import { classNames } from "../utils/class-names";
 import { generateURL } from "../utils/url";
-import {
-  CommentMention,
-  CommentNonInteractiveLink,
-  CommentNonInteractiveReaction,
-} from "./Comment";
 import { Avatar, type AvatarProps } from "./internal/Avatar";
 import { Button } from "./internal/Button";
 import { Dropdown, DropdownItem, DropdownTrigger } from "./internal/Dropdown";
+import {
+  generateInboxNotificationThreadContents,
+  INBOX_NOTIFICATION_THREAD_MAX_COMMENTS,
+  InboxNotificationComment,
+} from "./internal/InboxNotificationThread";
 import { List } from "./internal/List";
 import { Room } from "./internal/Room";
 import { Tooltip } from "./internal/Tooltip";
 import { User } from "./internal/User";
 
-const INBOX_NOTIFICATION_THREAD_MAX_COMMENTS = 3;
-
-type InboxNotificationThreadCommentsContents = {
-  type: "comments";
-  unread: boolean;
-  comments: CommentData[];
-  userIds: string[];
-  date: Date;
-};
-
-type InboxNotificationThreadMentionContents = {
-  type: "mention";
-  unread: boolean;
-  comments: CommentData[];
-  userIds: string[];
-  date: Date;
-};
-
-type InboxNotificationThreadContents =
-  | InboxNotificationThreadCommentsContents
-  | InboxNotificationThreadMentionContents;
-
-export type InboxNotificationKinds = {
+export type InboxNotificationKinds = Record<
+  string,
+  ComponentType<InboxNotificationCustomProps>
+> & {
   thread: ComponentType<InboxNotificationThreadProps>;
 };
 
-export type AddRefToComponents<T, R> = {
+type AddRefToComponents<T, R> = {
   [K in keyof T]: T[K] extends ComponentType<infer P>
     ? ComponentType<P & { ref: R }>
     : T[K];
@@ -96,7 +70,7 @@ interface InboxNotificationSharedProps {
 }
 
 export interface InboxNotificationProps
-  extends ComponentPropsWithoutRef<"a">,
+  extends Omit<ComponentPropsWithoutRef<"a">, "title">,
     InboxNotificationSharedProps {
   /**
    * The inbox notification to display.
@@ -125,9 +99,38 @@ export interface InboxNotificationThreadProps
   extends Omit<InboxNotificationProps, "kinds">,
     InboxNotificationSharedProps {
   /**
+   * The inbox notification to display.
+   */
+  inboxNotification: InboxNotificationThreadData;
+
+  /**
    * Whether to show the room name in the title.
    */
   showRoomName?: boolean;
+}
+
+export interface InboxNotificationCustomProps
+  extends Omit<InboxNotificationProps, "kinds">,
+    InboxNotificationSharedProps {
+  /**
+   * The inbox notification to display.
+   */
+  inboxNotification: InboxNotificationCustomData;
+
+  /**
+   * TODO: JSDoc
+   */
+  title?: ReactNode;
+
+  /**
+   * TODO: JSDoc
+   */
+  body?: ReactNode;
+
+  /**
+   * TODO: JSDoc
+   */
+  aside?: ReactNode;
 }
 
 interface InboxNotificationLayoutProps
@@ -143,12 +146,6 @@ interface InboxNotificationLayoutProps
 }
 
 type InboxNotificationAvatarProps = AvatarProps;
-
-interface InboxNotificationCommentProps extends ComponentProps<"div"> {
-  comment: CommentData;
-  showHeader?: boolean;
-  overrides?: Partial<GlobalOverrides & CommentOverrides>;
-}
 
 const InboxNotificationLayout = forwardRef<
   HTMLAnchorElement,
@@ -286,154 +283,6 @@ function InboxNotificationAvatar({
       {...props}
     />
   );
-}
-
-function InboxNotificationComment({
-  comment,
-  showHeader = true,
-  overrides,
-  className,
-  ...props
-}: InboxNotificationCommentProps) {
-  const $ = useOverrides(overrides);
-
-  return (
-    <div
-      className={classNames(
-        "lb-root lb-inbox-notification-comment lb-comment",
-        className
-      )}
-      {...props}
-    >
-      {showHeader && (
-        <div className="lb-comment-header">
-          <User className="lb-comment-author" userId={comment.userId} />
-        </div>
-      )}
-      <div className="lb-comment-content">
-        {comment.body ? (
-          <>
-            <CommentPrimitive.Body
-              className="lb-comment-body"
-              body={comment.body}
-              components={{
-                Mention: CommentMention,
-                Link: CommentNonInteractiveLink,
-              }}
-            />
-            {comment.reactions.length > 0 && (
-              <div className="lb-comment-reactions">
-                {comment.reactions.map((reaction) => (
-                  <CommentNonInteractiveReaction
-                    key={reaction.emoji}
-                    reaction={reaction}
-                    overrides={overrides}
-                    disabled
-                  />
-                ))}
-              </div>
-            )}
-          </>
-        ) : (
-          <div className="lb-comment-body">
-            <p className="lb-comment-deleted">{$.COMMENT_DELETED}</p>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-/**
- * Find the last comment with a mention for the given user ID,
- * unless the comment was created by the user themselves.
- */
-function findLastCommentWithMentionedId(
-  comments: CommentData[],
-  mentionedId: string
-) {
-  for (let i = comments.length - 1; i >= 0; i--) {
-    const comment = comments[i];
-
-    if (comment.userId === mentionedId) {
-      continue;
-    }
-
-    if (comment.body) {
-      const mentionedIds = getMentionedIdsFromCommentBody(comment.body);
-
-      if (mentionedIds.includes(mentionedId)) {
-        return comment;
-      }
-    }
-  }
-
-  return;
-}
-
-function getUserIdsFromComments(comments: CommentData[]) {
-  return Array.from(new Set(comments.map((comment) => comment.userId)));
-}
-
-function generateInboxNotificationThreadContents(
-  inboxNotification: InboxNotificationThreadData,
-  thread: ThreadData<BaseMetadata>,
-  userId: string
-): InboxNotificationThreadContents {
-  const unreadComments = thread.comments.filter((comment) => {
-    if (!comment.body) {
-      return false;
-    }
-
-    return inboxNotification.readAt
-      ? comment.createdAt > inboxNotification.readAt &&
-          comment.createdAt <= inboxNotification.notifiedAt
-      : comment.createdAt <= inboxNotification.notifiedAt;
-  });
-
-  // If the thread is read, show the last comments.
-  if (unreadComments.length === 0) {
-    const lastComments = thread.comments
-      .filter((comment) => comment.body)
-      .slice(-INBOX_NOTIFICATION_THREAD_MAX_COMMENTS);
-
-    return {
-      type: "comments",
-      unread: false,
-      comments: lastComments,
-      userIds: getUserIdsFromComments(lastComments),
-      date: inboxNotification.notifiedAt,
-    };
-  }
-
-  const commentWithMention = findLastCommentWithMentionedId(
-    unreadComments,
-    userId
-  );
-
-  // If the thread contains one or more mentions for the current user, show the last comment with a mention.
-  if (commentWithMention) {
-    return {
-      type: "mention",
-      unread: true,
-      comments: [commentWithMention],
-      userIds: [commentWithMention.userId],
-      date: commentWithMention.createdAt,
-    };
-  }
-
-  const lastUnreadComments = unreadComments.slice(
-    -INBOX_NOTIFICATION_THREAD_MAX_COMMENTS
-  );
-
-  // Otherwise, show the last unread comments.
-  return {
-    type: "comments",
-    unread: true,
-    comments: lastUnreadComments,
-    userIds: getUserIdsFromComments(unreadComments),
-    date: inboxNotification.notifiedAt,
-  };
 }
 
 /**
@@ -585,8 +434,10 @@ const InboxNotificationThread = forwardRef<
   }
 );
 
+// TODO: Remove `as any`
 const defaultInboxNotificationKinds: InboxNotificationKinds = {
-  thread: InboxNotificationThread,
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+  thread: InboxNotificationThread as any,
 };
 
 /**
@@ -606,7 +457,7 @@ const defaultInboxNotificationKinds: InboxNotificationKinds = {
 export const InboxNotification = Object.assign(
   forwardRef<HTMLAnchorElement, InboxNotificationProps>(
     ({ inboxNotification, kinds, ...props }, forwardedRef) => {
-      const { thread: InboxNotificationThread } = useMemo(
+      const { thread: InboxNotificationThread, ...resolvedKinds } = useMemo(
         () =>
           ({
             ...defaultInboxNotificationKinds,
@@ -619,17 +470,34 @@ export const InboxNotification = Object.assign(
         case "thread":
           return (
             <InboxNotificationThread
-              inboxNotification={inboxNotification}
+              inboxNotification={
+                inboxNotification as InboxNotificationThreadData
+              }
               {...props}
               ref={forwardedRef}
             />
           );
 
-        default:
-          return assertNever(
-            inboxNotification.kind,
-            "Unexpected inbox notification kind"
+        default: {
+          const InboxNotificationCustom = resolvedKinds[inboxNotification.kind];
+
+          if (!InboxNotificationCustom) {
+            // TODO: Don't render null, render an empty notification instead with just the time and dropdown.
+            // TODO: Warn in the console that this kind wasn't handled.
+
+            return null;
+          }
+
+          return (
+            <InboxNotificationCustom
+              inboxNotification={
+                inboxNotification as InboxNotificationCustomData
+              }
+              {...props}
+              ref={forwardedRef}
+            />
           );
+        }
       }
     }
   ),
